+++ logic/include/LogicSystem.h
@@ -21,7 +21,7 @@ struct LogicSystem : public ObjectSystem, public LogicAllocator {
 	
 	virtual LogicComponent* createComponent() = 0;
 	
-	virtual void execute(ObjectManager* objectManager, long delta) = 0;
+	virtual void execute(ObjectManager* objectManager, double delta) = 0;
 	virtual std::uint8_t getSysId() const = 0;	
 	
 	virtual void setLog(Log* log, const char* target = 0) = 0;
