+++ rendering/include/RenderingSystem.h
@@ -33,7 +33,7 @@ struct RenderingSystem : public RenderingAllocator, public ObjectSystem {
 	virtual void destroyGlobalPrimitive(Primitive* prmi) = 0;
 
 	virtual std::uint8_t getSysId() const = 0;
-	virtual void execute(ObjectManager* objectManager, long delta) = 0;
+	virtual void execute(ObjectManager* objectManager, double delta) = 0;
 
 	virtual void setLog(Log* log, const char* target = 0) = 0;
